Struct flowrcli::cli::connections::ClientConnection

pub struct ClientConnection {
    requester: Socket,
}

ClientConnection stores information related to the connection from a client to the Coordinator and is used each time a message is to be sent or received.

Fields
requester: Socket
Implementations

impl ClientConnection

pub fn new(coordinator_address: &str) -> Result<Self>

Create a new connection between client and Coordinator

pub fn receive<CM>(&self) -> Result<CM>where CM: From<String> + Display,

Receive a CoordinatorMessage from the Coordinator

pub fn send<CM>(&self, message: CM) -> Result<()>where CM: Into<String> + Display,

Send a CoordinatorMessage to the Coordinator